@@ -5,13 +5,13 @@
   "homepage": ".",
   "dependencies": {
     "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^11.2.7",
+    "@testing-library/react": "^16.3.2",
     "@testing-library/user-event": "^12.8.3",
-    "react": "^17.0.2",
+    "react": "^19.2.4",
     "react-charts": "^2.0.0-beta.7",
-    "react-dom": "^17.0.2",
+    "react-dom": "^19.2.4",
     "react-lazy-load-image-component": "^1.5.1",
-    "react-redux": "^7.2.4",
+    "react-redux": "^9.2.0",
     "react-router": "^5.1.2",
     "react-router-dom": "^5.2.0",
     "recharts": "^2.0.9",
