@@ -5,19 +5,19 @@
   "homepage": ".",
   "dependencies": {
     "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^11.2.7",
+    "@testing-library/react": "^16.3.2",
     "@testing-library/user-event": "^13.1.9",
     "axios": "^0.21.1",
     "dateformat": "^4.5.1",
     "lodash": "^4.17.21",
     "moment": "^2.29.1",
-    "react": "^17.0.2",
+    "react": "^19.2.4",
     "react-big-calendar": "^0.33.5",
     "react-device-detect": "^1.17.0",
-    "react-dom": "^17.0.2",
+    "react-dom": "^19.2.4",
     "react-lazy-load-image-component": "^1.5.1",
     "react-player": "^2.9.0",
-    "react-redux": "^7.2.4",
+    "react-redux": "^9.2.0",
     "react-router": "^5.2.0",
     "react-router-dom": "^5.1.2",
     "redux": "^4.1.0",
