@@ -1,5 +1,5 @@
 import React from "react";
-import ReactDOM from "react-dom";
+import { createRoot } from "react-dom/client";
 import App from "./App";
 import * as serviceWorker from "./serviceWorker";
 import { initStore, store } from "./data/store";
@@ -8,11 +8,12 @@ import "./styles/main.scss";
 
 const startApp = () => {
   initStore();
-  ReactDOM.render(
+  const container = document.getElementById("root");
+  const root = createRoot(container);
+  root.render(
     <Provider store={store}>
       <App />
-    </Provider>,
-    document.getElementById("root")
+    </Provider>
   );
 };
 
